@@ -2,12 +2,9 @@
 // The .NET Foundation licenses this file to you under the MS-PL license.
 // See the LICENSE file in the project root for more information.
 
-namespace Playground.Core.Models
-{
-    public class SampleModel
-    {
-        public string Message { get; set; }
+namespace Playground.Core.Models;
 
-        public decimal Value { get; set; }
-    }
+public record SampleModel(string Message, decimal Value)
+{
+    public override string ToString() => $"Message: {Message}, Value: {Value}";
 }
