@@ -2,13 +2,6 @@
 // The .NET Foundation licenses this file to you under the MS-PL license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using MvvmCross.Base;
-using MvvmCross.Logging;
-using MvvmCross.Tests;
 using MvvmCross.ViewModels;
 using MvvmCross.Views;
 
@@ -17,13 +10,12 @@ namespace MvvmCross.UnitTest.Mocks.Dispatchers
     public class NavigationMockDispatcher
         : IMvxViewDispatcher
     {
-        public readonly List<MvxViewModelRequest> Requests = new List<MvxViewModelRequest>();
-        public readonly List<MvxPresentationHint> Hints = new List<MvxPresentationHint>();
+        public readonly List<MvxViewModelRequest> Requests = [];
+        public readonly List<MvxPresentationHint> Hints = [];
 
         public bool IsOnMainThread => true;
 
-        public virtual bool RequestMainThreadAction(Action action,
-                                                    bool maskExceptions = true)
+        public bool RequestMainThreadAction(Action action, bool maskExceptions = true)
         {
             try
             {
@@ -39,13 +31,13 @@ namespace MvvmCross.UnitTest.Mocks.Dispatchers
             }
         }
 
-        public virtual Task<bool> ShowViewModel(MvxViewModelRequest request)
+        public Task<bool> ShowViewModel(MvxViewModelRequest request)
         {
             Requests.Add(request);
             return Task.FromResult(true);
         }
 
-        public virtual Task<bool> ChangePresentation(MvxPresentationHint hint)
+        public Task<bool> ChangePresentation(MvxPresentationHint hint)
         {
             Hints.Add(hint);
             return Task.FromResult(true);
