@@ -611,7 +611,8 @@ public class MvxMultiWindowViewPresenter
 
         static Task TryEnqueue(DispatcherQueue dispatcher, Action function, DispatcherQueuePriority priority)
         {
-            var taskCompletionSource = new TaskCompletionSource<object?>();
+            var taskCompletionSource =
+                new TaskCompletionSource<object?>(TaskCreationOptions.RunContinuationsAsynchronously);
 
             if (!dispatcher.TryEnqueue(
                     priority,
@@ -702,14 +703,13 @@ public class MvxMultiWindowViewPresenter
             if (page is IMvxNeedWindow needWindow2 && !needWindow2.CanClose())
             {
                 e.Cancel = true;
-                return;
             }
 
-            CloseWindow(newWindow);
         };
 
         newWindow.Closed += (_, _) =>
         {
+            CloseWindow(newWindow);
             page.DisposeIfDisposable();
 
             lock (_windowInformationLock)
