@@ -2,10 +2,8 @@
 // The .NET Foundation licenses this file to you under the MS-PL license.
 // See the LICENSE file in the project root for more information.
 
-using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using MvvmCross.Commands;
-using MvvmCross.Logging;
 using MvvmCross.Navigation;
 using MvvmCross.ViewModels;
 using Playground.Core.Models;
@@ -17,14 +15,14 @@ namespace Playground.Core.ViewModels
         public string BrokenTextValue { get => _brokenTextValue; set => SetProperty(ref _brokenTextValue, value); }
         public string AnotherBrokenTextValue { get => _anotherBrokenTextValue; set => SetProperty(ref _anotherBrokenTextValue, value); }
 
-        private SampleModel _parameter;
         private string _brokenTextValue;
         private string _anotherBrokenTextValue;
 
-        public ChildViewModel(ILoggerFactory logProvider, IMvxNavigationService navigationService)
+        public ChildViewModel(
+            ILoggerFactory logProvider, IMvxNavigationService navigationService)
             : base(logProvider, navigationService)
         {
-            CloseCommand = new MvxAsyncCommand(() => NavigationService.Close(this));
+            CloseCommand = new MvxAsyncCommand(DoCloseCommand);
 
             ShowSecondChildCommand = new MvxAsyncCommand(() => NavigationService.Navigate<SecondChildViewModel>());
 
@@ -33,17 +31,17 @@ namespace Playground.Core.ViewModels
             PropertyChanged += ChildViewModel_PropertyChanged;
         }
 
+        private Task DoCloseCommand()
+        {
+            return NavigationService.Close(this);
+        }
+
         private void ChildViewModel_PropertyChanged(object sender, System.ComponentModel.PropertyChangedEventArgs e)
         {
             // Demonstrates that exceptions can be raised on property changed but are swallowed by default to 
             // protect the app from crashing
             if (e.PropertyName == nameof(BrokenTextValue))
                 throw new System.NotImplementedException();
-        }
-
-        public override void Prepare(SampleModel parameter)
-        {
-            _parameter = parameter;
         }
 
         public override async System.Threading.Tasks.Task Initialize()
@@ -63,6 +61,10 @@ namespace Playground.Core.ViewModels
         public IMvxAsyncCommand ShowSecondChildCommand { get; private set; }
 
         public IMvxAsyncCommand ShowRootCommand { get; private set; }
+
+        public override void Prepare(SampleModel parameter)
+        {
+        }
 
         public override void ViewAppeared()
         {
