@@ -1,15 +1,12 @@
-using UIKit;
+namespace Playground.iOS;
 
-namespace Playground.iOS
+public class Application
 {
-    public class Application
+    // This is the main entry point of the application.
+    private static void Main(string[] args)
     {
-        // This is the main entry point of the application.
-        private static void Main(string[] args)
-        {
-            // if you want to use a different Application Delegate class from "AppDelegate"
-            // you can specify it here.
-            UIApplication.Main(args, null, "AppDelegate");
-        }
+        // if you want to use a different Application Delegate class from "AppDelegate"
+        // you can specify it here.
+        UIApplication.Main(args, null, "AppDelegate");
     }
 }
