@@ -2,14 +2,12 @@
 // The .NET Foundation licenses this file to you under the MS-PL license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
-using Moq;
 using MvvmCross.Exceptions;
 using MvvmCross.Navigation.EventArguments;
-using MvvmCross.Tests;
 using MvvmCross.UnitTest.Mocks.TestViewModels;
 using MvvmCross.ViewModels;
+using NSubstitute;
+using NSubstitute.ExceptionExtensions;
 using Xunit;
 
 namespace MvvmCross.UnitTest.ViewModels
@@ -46,19 +44,23 @@ namespace MvvmCross.UnitTest.ViewModels
 
             IMvxViewModel outViewModel = new Test2ViewModel();
 
-            var mockLocator = new Mock<IMvxViewModelLocator>();
-            mockLocator.Setup(
-                m => m.Load(It.IsAny<Type>(), It.IsAny<IMvxBundle>(), It.IsAny<IMvxBundle>(), It.IsAny<IMvxNavigateEventArgs>()))
-                       .Returns(() => outViewModel);
+            var mockLocator = Substitute.For<IMvxViewModelLocator>();
+            mockLocator.Load(
+                    Arg.Any<Type>(),
+                    Arg.Any<IMvxBundle>(),
+                    Arg.Any<IMvxBundle>(),
+                    Arg.Any<IMvxNavigateEventArgs>())
+                       .Returns(outViewModel);
 
-            var mockCollection = new Mock<IMvxViewModelLocatorCollection>();
-            mockCollection.Setup(m => m.FindViewModelLocator(It.IsAny<MvxViewModelRequest>()))
-                          .Returns(() => mockLocator.Object);
+            var mockCollection = Substitute.For<IMvxViewModelLocatorCollection>();
+            mockCollection
+                .FindViewModelLocator(Arg.Any<MvxViewModelRequest>())
+                .Returns(mockLocator);
 
             var parameters = new Dictionary<string, string> { { "foo", "bar" } };
             var request = new MvxViewModelRequest<Test2ViewModel>(new MvxBundle(parameters), null);
             var state = new MvxBundle();
-            var loader = new MvxViewModelLoader(mockCollection.Object);
+            var loader = new MvxViewModelLoader(mockCollection);
             var args = new MvxNavigateEventArgs(NavigationMode.Show);
             var viewModel = loader.LoadViewModel(request, state, args);
 
@@ -70,19 +72,22 @@ namespace MvvmCross.UnitTest.ViewModels
         {
             _fixture.ClearAll();
 
-            var mockLocator = new Mock<IMvxViewModelLocator>();
-            mockLocator.Setup(
-                m => m.Load(It.IsAny<Type>(), It.IsAny<IMvxBundle>(), It.IsAny<IMvxBundle>(), It.IsAny<IMvxNavigateEventArgs>()))
-                       .Throws<MvxException>();
+            var mockLocator = Substitute.For<IMvxViewModelLocator>();
+            mockLocator.Load(
+                    Arg.Any<Type>(),
+                    Arg.Any<IMvxBundle>(),
+                    Arg.Any<IMvxBundle>(),
+                    Arg.Any<IMvxNavigateEventArgs>())
+                .Throws<MvxException>();
 
-            var mockCollection = new Mock<IMvxViewModelLocatorCollection>();
-            mockCollection.Setup(m => m.FindViewModelLocator(It.IsAny<MvxViewModelRequest>()))
-                          .Returns(() => mockLocator.Object);
+            var mockCollection = Substitute.For<IMvxViewModelLocatorCollection>();
+            mockCollection.FindViewModelLocator(Arg.Any<MvxViewModelRequest>())
+                          .Returns(mockLocator);
 
             var parameters = new Dictionary<string, string> { { "foo", "bar" } };
             var request = new MvxViewModelRequest<Test2ViewModel>(new MvxBundle(parameters), null);
             var state = new MvxBundle();
-            var loader = new MvxViewModelLoader(mockCollection.Object);
+            var loader = new MvxViewModelLoader(mockCollection);
             var args = new MvxNavigateEventArgs(NavigationMode.Show);
             Assert.Throws<MvxException>(() =>
             {
@@ -95,14 +100,14 @@ namespace MvvmCross.UnitTest.ViewModels
         {
             _fixture.ClearAll();
 
-            var mockCollection = new Mock<IMvxViewModelLocatorCollection>();
-            mockCollection.Setup(m => m.FindViewModelLocator(It.IsAny<MvxViewModelRequest>()))
-                          .Returns(() => null);
+            var mockCollection = Substitute.For<IMvxViewModelLocatorCollection>();
+            mockCollection.FindViewModelLocator(Arg.Any<MvxViewModelRequest>())
+                .Returns((IMvxViewModelLocator?)null);
 
             var parameters = new Dictionary<string, string> { { "foo", "bar" } };
             var request = new MvxViewModelRequest<Test2ViewModel>(new MvxBundle(parameters), null);
             var state = new MvxBundle();
-            var loader = new MvxViewModelLoader(mockCollection.Object);
+            var loader = new MvxViewModelLoader(mockCollection);
             var args = new MvxNavigateEventArgs(NavigationMode.Show);
             Assert.Throws<MvxException>(() =>
             {
